314: Umsetzung von get_component und Finalisierung ist erfolgt
